+++ packages/tsrx-react/src/transform.js
@@ -4,7 +4,14 @@
 import { walk } from 'zimmerframe';
 import { print } from 'esrap';
 import tsx from 'esrap/languages/tsx';
-import { renderStylesheets, setLocation } from '@tsrx/core';
+import {
+	renderStylesheets,
+	setLocation,
+	applyLazyTransforms as apply_lazy_transforms,
+	collectLazyBindingsFromComponent as collect_lazy_bindings_from_component,
+	preallocateLazyIds as preallocate_lazy_ids,
+	replaceLazyParams as replace_lazy_params,
+} from '@tsrx/core';
 
 /**
  * @typedef {{
@@ -173,713 +180,6 @@ export function transform(ast, source, filename) {
 			: null;
 
 	return { ast: final_program, code: result.code, map: result.map, css };
-}
-
-// --- Lazy destructuring support ---
-
-/**
- * Generate a unique lazy identifier name for a lazy destructuring pattern.
- * @param {TransformContext} transform_context
- * @returns {string}
- */
-function generate_lazy_id(transform_context) {
-	return `__lazy${transform_context.lazy_next_id++}`;
-}
-
-/**
- * Collect lazy bindings from a destructuring pattern.
- * For `&{name, age}`, maps `name` → `source.name`, `age` → `source.age`.
- * For `&[a, b]`, maps `a` → `source[0]`, `b` → `source[1]`.
- * Handles nested AssignmentPattern (default values) and RestElement.
- *
- * @param {any} pattern - The ObjectPattern or ArrayPattern with lazy: true
- * @param {string} source_name - The generated identifier name for the source
- * @param {Map<string, LazyBinding>} lazy_bindings - Map to populate
- */
-function collect_lazy_bindings(pattern, source_name, lazy_bindings) {
-	if (pattern.type === 'ObjectPattern') {
-		for (const prop of pattern.properties || []) {
-			if (prop.type === 'RestElement') {
-				// Rest element in object pattern — skip for now (complex to transform)
-				continue;
-			}
-			const value = prop.value;
-			const actual = value.type === 'AssignmentPattern' ? value.left : value;
-			if (actual.type === 'Identifier') {
-				const key = prop.key;
-				const computed = prop.computed || key.type !== 'Identifier';
-				lazy_bindings.set(actual.name, {
-					source_name,
-					read: () => ({
-						type: 'MemberExpression',
-						object: create_generated_identifier(source_name),
-						property: computed
-							? { ...key }
-							: { type: 'Identifier', name: key.name, metadata: { path: [] } },
-						computed,
-						optional: false,
-						metadata: { path: [] },
-					}),
-				});
-			}
-		}
-	} else if (pattern.type === 'ArrayPattern') {
-		for (let i = 0; i < (pattern.elements || []).length; i++) {
-			const element = pattern.elements[i];
-			if (!element) continue;
-			if (element.type === 'RestElement') {
-				// Rest element in array pattern — skip for now
-				continue;
-			}
-			const actual = element.type === 'AssignmentPattern' ? element.left : element;
-			if (actual.type === 'Identifier') {
-				const index = i;
-				lazy_bindings.set(actual.name, {
-					source_name,
-					read: () => ({
-						type: 'MemberExpression',
-						object: create_generated_identifier(source_name),
-						property: { type: 'Literal', value: index, raw: String(index), metadata: { path: [] } },
-						computed: true,
-						optional: false,
-						metadata: { path: [] },
-					}),
-				});
-			}
-		}
-	}
-}
-
-/**
- * Collect lazy bindings from component params and body variable declarations
- * WITHOUT modifying any AST nodes. Returns a map of binding name → accessor info.
- * Stores the generated identifier name on the pattern's metadata for later replacement.
- *
- * @param {any[]} params - Component params (metadata annotated, not structurally mutated)
- * @param {any[]} body - Component body (metadata annotated, not structurally mutated)
- * @param {TransformContext} transform_context
- * @returns {Map<string, LazyBinding>}
- */
-function collect_lazy_bindings_from_component(params, body, transform_context) {
-	/** @type {Map<string, LazyBinding>} */
-	const lazy_bindings = new Map();
-
-	// Collect from lazy params
-	for (const param of params) {
-		const pattern = param.type === 'AssignmentPattern' ? param.left : param;
-
-		if ((pattern.type === 'ObjectPattern' || pattern.type === 'ArrayPattern') && pattern.lazy) {
-			const lazy_name = pattern.metadata?.lazy_id || generate_lazy_id(transform_context);
-			if (!pattern.metadata?.lazy_id) {
-				pattern.metadata = { ...pattern.metadata, lazy_id: lazy_name };
-			}
-			collect_lazy_bindings(pattern, lazy_name, lazy_bindings);
-		}
-	}
-
-	// Collect from lazy variable declarations in body
-	for (const statement of body) {
-		if (statement.type !== 'VariableDeclaration') continue;
-
-		for (const declarator of statement.declarations || []) {
-			const pattern = declarator.id;
-			if ((pattern.type === 'ObjectPattern' || pattern.type === 'ArrayPattern') && pattern.lazy) {
-				const lazy_name = pattern.metadata?.lazy_id || generate_lazy_id(transform_context);
-				if (!pattern.metadata?.lazy_id) {
-					pattern.metadata = { ...pattern.metadata, lazy_id: lazy_name };
-				}
-				collect_lazy_bindings(pattern, lazy_name, lazy_bindings);
-			}
-		}
-	}
-
-	// Collect from standalone lazy assignment statements: `&[x] = expr;`
-	collect_lazy_bindings_from_statements(body, lazy_bindings);
-
-	return lazy_bindings;
-}
-
-/**
- * Collect lazy bindings from statements at the top level of a block.
- * Reads already-allocated `lazy_id` values stored on pattern metadata.
- * Handles both `let &[x] = ...` VariableDeclarations and statement-level
- * `&[x] = expr;` ExpressionStatement assignments.
- *
- * @param {any[]} statements
- * @param {Map<string, LazyBinding>} lazy_bindings
- */
-function collect_lazy_bindings_from_statements(statements, lazy_bindings) {
-	for (const stmt of statements || []) {
-		if (stmt.type === 'VariableDeclaration') {
-			for (const declarator of stmt.declarations || []) {
-				const pattern = declarator.id;
-				if (
-					(pattern?.type === 'ObjectPattern' || pattern?.type === 'ArrayPattern') &&
-					pattern.lazy &&
-					pattern.metadata?.lazy_id &&
-					!lazy_bindings_contains(lazy_bindings, pattern)
-				) {
-					collect_lazy_bindings(pattern, pattern.metadata.lazy_id, lazy_bindings);
-				}
-			}
-		} else if (
-			stmt.type === 'ExpressionStatement' &&
-			stmt.expression?.type === 'AssignmentExpression' &&
-			stmt.expression.operator === '=' &&
-			(stmt.expression.left?.type === 'ObjectPattern' ||
-				stmt.expression.left?.type === 'ArrayPattern') &&
-			stmt.expression.left.lazy &&
-			stmt.expression.left.metadata?.lazy_id
-		) {
-			collect_lazy_bindings(
-				stmt.expression.left,
-				stmt.expression.left.metadata.lazy_id,
-				lazy_bindings,
-			);
-		}
-	}
-}
-
-/**
- * Return true if any of the names in the pattern already exist in the map.
- * @param {Map<string, LazyBinding>} lazy_bindings
- * @param {any} pattern
- * @returns {boolean}
- */
-function lazy_bindings_contains(lazy_bindings, pattern) {
-	if (pattern.type === 'ObjectPattern') {
-		for (const prop of pattern.properties || []) {
-			if (prop.type === 'RestElement') continue;
-			const value = prop.value;
-			const actual = value?.type === 'AssignmentPattern' ? value.left : value;
-			if (actual?.type === 'Identifier' && lazy_bindings.has(actual.name)) return true;
-		}
-	} else if (pattern.type === 'ArrayPattern') {
-		for (const element of pattern.elements || []) {
-			if (!element || element.type === 'RestElement') continue;
-			const actual = element.type === 'AssignmentPattern' ? element.left : element;
-			if (actual?.type === 'Identifier' && lazy_bindings.has(actual.name)) return true;
-		}
-	}
-	return false;
-}
-
-/**
- * Walk the AST and pre-allocate `lazy_id` metadata on every lazy destructuring
- * pattern: function/component params, variable declarator ids, and statement-level
- * assignment LHS. Idempotent: skips patterns that already have a `lazy_id`.
- * Runs before the main transform so every descent knows the final generated name.
- *
- * @param {any} root
- * @param {TransformContext} transform_context
- */
-function preallocate_lazy_ids(root, transform_context) {
-	/** @param {any} pattern */
-	const assign_id = (pattern) => {
-		if (
-			(pattern?.type === 'ObjectPattern' || pattern?.type === 'ArrayPattern') &&
-			pattern.lazy &&
-			!pattern.metadata?.lazy_id
-		) {
-			pattern.metadata = {
-				...pattern.metadata,
-				lazy_id: generate_lazy_id(transform_context),
-			};
-		}
-	};
-
-	/** @param {any} node */
-	const visit = (node) => {
-		if (!node || typeof node !== 'object') return;
-		if (Array.isArray(node)) {
-			for (const child of node) visit(child);
-			return;
-		}
-
-		if (
-			node.type === 'FunctionDeclaration' ||
-			node.type === 'FunctionExpression' ||
-			node.type === 'ArrowFunctionExpression' ||
-			node.type === 'Component'
-		) {
-			for (const param of node.params || []) {
-				assign_id(param?.type === 'AssignmentPattern' ? param.left : param);
-			}
-		}
-
-		if (node.type === 'VariableDeclarator') {
-			assign_id(node.id);
-		}
-
-		if (
-			node.type === 'ExpressionStatement' &&
-			node.expression?.type === 'AssignmentExpression' &&
-			node.expression.operator === '='
-		) {
-			assign_id(node.expression.left);
-		}
-
-		for (const key of Object.keys(node)) {
-			if (key === 'loc' || key === 'start' || key === 'end' || key === 'metadata') continue;
-			visit(node[key]);
-		}
-	};
-
-	visit(root);
-}
-
-/**
- * Walk an AST node tree and replace identifier references that match lazy bindings
- * with their corresponding member expressions (e.g., `name` → `__lazy0.name`).
- * Also handles AssignmentExpression and UpdateExpression targets.
- *
- * @param {any} node - The AST node to walk
- * @param {Map<string, LazyBinding>} lazy_bindings - Map of lazy binding names
- * @returns {any}
- */
-function apply_lazy_transforms(node, lazy_bindings) {
-	if (!node || typeof node !== 'object') return node;
-	if (Array.isArray(node)) return node.map((child) => apply_lazy_transforms(child, lazy_bindings));
-
-	// Don't recurse into nested function declarations (helper components have their own scope)
-	if (
-		node.type === 'FunctionDeclaration' ||
-		node.type === 'FunctionExpression' ||
-		node.type === 'ArrowFunctionExpression'
-	) {
-		// Transform default parameter values (e.g. (step = count) => ...) with the
-		// outer lazy_bindings, since defaults are evaluated in the outer scope.
-		let params_changed = false;
-		const new_params = (node.params || []).map((/** @type {any} */ param) => {
-			const transformed = transform_param_defaults(param, lazy_bindings);
-			if (transformed !== param) params_changed = true;
-			return transformed;
-		});
-
-		// Check if any params shadow a lazy binding — if so, exclude those names
-		/** @type {Set<string>} */
-		const shadowed = new Set();
-		for (const param of node.params || []) {
-			collect_shadowed_names(param, lazy_bindings, shadowed);
-		}
-
-		const outer_minus_shadow =
-			shadowed.size > 0 ? remove_shadowed(lazy_bindings, shadowed) : lazy_bindings;
-
-		// Collect this function's own lazy param bindings (pre-allocated lazy_ids).
-		/** @type {Map<string, LazyBinding>} */
-		const own_bindings = new Map();
-		let had_lazy_param = false;
-		for (const param of node.params || []) {
-			const pattern = param?.type === 'AssignmentPattern' ? param.left : param;
-			if (
-				(pattern?.type === 'ObjectPattern' || pattern?.type === 'ArrayPattern') &&
-				pattern.lazy &&
-				pattern.metadata?.lazy_id
-			) {
-				had_lazy_param = true;
-				collect_lazy_bindings(pattern, pattern.metadata.lazy_id, own_bindings);
-			}
-		}
-
-		// Own bindings override any outer binding with the same name.
-		const inner_bindings =
-			own_bindings.size > 0
-				? new Map([...outer_minus_shadow, ...own_bindings])
-				: outer_minus_shadow;
-
-		if (inner_bindings.size === 0 && !params_changed && !had_lazy_param) return node;
-
-		const new_body =
-			inner_bindings.size > 0 ? apply_lazy_transforms(node.body, inner_bindings) : node.body;
-
-		const final_params_src = params_changed ? new_params : node.params;
-		const final_params = had_lazy_param ? replace_lazy_params(final_params_src) : final_params_src;
-
-		if (new_body !== node.body || final_params !== node.params) {
-			return {
-				...node,
-				params: final_params,
-				body: new_body,
-			};
-		}
-		return node;
-	}
-
-	// Handle block-scoped variable shadowing (const/let/var that shadows a lazy name)
-	if (node.type === 'BlockStatement' || node.type === 'Program') {
-		const block_bindings = collect_block_shadowed_names(node.body, lazy_bindings);
-		const after_shadow =
-			block_bindings.size > 0 ? remove_shadowed(lazy_bindings, block_bindings) : lazy_bindings;
-
-		// Collect new lazy bindings declared in this block (lazy VariableDeclarations
-		// and statement-level `&[x] = expr;` assignments). These become in-scope for
-		// the block's statements.
-		/** @type {Map<string, LazyBinding>} */
-		const block_lazy = new Map();
-		collect_lazy_bindings_from_statements(node.body, block_lazy);
-
-		const effective_bindings =
-			block_lazy.size > 0 ? new Map([...after_shadow, ...block_lazy]) : after_shadow;
-
-		// Always descend — nested functions may have their own lazy params that
-		// need transforming even when no bindings are in scope at this level.
-		let changed = false;
-		const new_body = node.body.map((/** @type {any} */ stmt) => {
-			const transformed = apply_lazy_transforms(stmt, effective_bindings);
-			if (transformed !== stmt) changed = true;
-			return transformed;
-		});
-		return changed ? { ...node, body: new_body } : node;
-	}
-
-	// Handle catch clause parameter shadowing
-	if (node.type === 'CatchClause') {
-		/** @type {Set<string>} */
-		const shadowed = new Set();
-		if (node.param) collect_shadowed_names(node.param, lazy_bindings, shadowed);
-		const effective_bindings =
-			shadowed.size > 0 ? remove_shadowed(lazy_bindings, shadowed) : lazy_bindings;
-		const new_body = apply_lazy_transforms(node.body, effective_bindings);
-		if (new_body !== node.body) return { ...node, body: new_body };
-		return node;
-	}
-
-	// Handle for-loop variable shadowing
-	if (node.type === 'ForStatement') {
-		/** @type {Set<string>} */
-		const shadowed = new Set();
-		if (node.init?.type === 'VariableDeclaration') {
-			for (const decl of node.init.declarations) {
-				if (decl.id) collect_shadowed_names(decl.id, lazy_bindings, shadowed);
-			}
-		}
-		const effective_bindings =
-			shadowed.size > 0 ? remove_shadowed(lazy_bindings, shadowed) : lazy_bindings;
-		let changed = false;
-		const new_init = apply_lazy_transforms(node.init, effective_bindings);
-		if (new_init !== node.init) changed = true;
-		const new_test = apply_lazy_transforms(node.test, effective_bindings);
-		if (new_test !== node.test) changed = true;
-		const new_update = apply_lazy_transforms(node.update, effective_bindings);
-		if (new_update !== node.update) changed = true;
-		const new_body = apply_lazy_transforms(node.body, effective_bindings);
-		if (new_body !== node.body) changed = true;
-		return changed
-			? { ...node, init: new_init, test: new_test, update: new_update, body: new_body }
-			: node;
-	}
-
-	if (node.type === 'ForOfStatement' || node.type === 'ForInStatement') {
-		/** @type {Set<string>} */
-		const shadowed = new Set();
-		if (node.left?.type === 'VariableDeclaration') {
-			for (const decl of node.left.declarations) {
-				if (decl.id) collect_shadowed_names(decl.id, lazy_bindings, shadowed);
-			}
-		}
-		const effective_bindings =
-			shadowed.size > 0 ? remove_shadowed(lazy_bindings, shadowed) : lazy_bindings;
-		let changed = false;
-		const new_right = apply_lazy_transforms(node.right, lazy_bindings);
-		if (new_right !== node.right) changed = true;
-		const new_body = apply_lazy_transforms(node.body, effective_bindings);
-		if (new_body !== node.body) changed = true;
-		return changed ? { ...node, right: new_right, body: new_body } : node;
-	}
-
-	// Handle switch-case variable shadowing (const/let inside case consequent arrays)
-	if (node.type === 'SwitchStatement') {
-		let changed = false;
-		const new_discriminant = apply_lazy_transforms(node.discriminant, lazy_bindings);
-		if (new_discriminant !== node.discriminant) changed = true;
-		const new_cases = node.cases.map((/** @type {any} */ switch_case) => {
-			const case_bindings = collect_block_shadowed_names(switch_case.consequent, lazy_bindings);
-			const effective_bindings =
-				case_bindings.size > 0 ? remove_shadowed(lazy_bindings, case_bindings) : lazy_bindings;
-			let case_changed = false;
-			const new_test = switch_case.test
-				? apply_lazy_transforms(switch_case.test, lazy_bindings)
-				: null;
-			if (new_test !== switch_case.test) case_changed = true;
-			const new_consequent = switch_case.consequent.map((/** @type {any} */ stmt) => {
-				const transformed = apply_lazy_transforms(stmt, effective_bindings);
-				if (transformed !== stmt) case_changed = true;
-				return transformed;
-			});
-			if (case_changed) {
-				changed = true;
-				return { ...switch_case, test: new_test, consequent: new_consequent };
-			}
-			return switch_case;
-		});
-		return changed ? { ...node, discriminant: new_discriminant, cases: new_cases } : node;
-	}
-
-	// Handle standalone lazy destructuring assignment: `&[data] = track(0);`
-	// becomes `const __lazy0 = track(0);`. The individual name bindings are
-	// already in scope via the enclosing BlockStatement handler.
-	if (
-		node.type === 'ExpressionStatement' &&
-		node.expression?.type === 'AssignmentExpression' &&
-		node.expression.operator === '=' &&
-		(node.expression.left?.type === 'ObjectPattern' ||
-			node.expression.left?.type === 'ArrayPattern') &&
-		node.expression.left.lazy &&
-		node.expression.left.metadata?.lazy_id
-	) {
-		const pattern = node.expression.left;
-		const lazy_id = create_generated_identifier(pattern.metadata.lazy_id);
-		if (pattern.typeAnnotation) {
-			lazy_id.typeAnnotation = pattern.typeAnnotation;
-		}
-		const init = apply_lazy_transforms(node.expression.right, lazy_bindings);
-		return /** @type {any} */ ({
-			type: 'VariableDeclaration',
-			kind: 'const',
-			declarations: [
-				{
-					type: 'VariableDeclarator',
-					id: lazy_id,
-					init,
-					metadata: { path: [] },
-				},
-			],
-			metadata: { path: [] },
-		});
-	}
-
-	// Handle assignment: `name = value` → `__lazy0.name = value`
-	if (node.type === 'AssignmentExpression' && node.left.type === 'Identifier') {
-		const binding = lazy_bindings.get(node.left.name);
-		if (binding) {
-			return {
-				...node,
-				left: binding.read(),
-				right: apply_lazy_transforms(node.right, lazy_bindings),
-			};
-		}
-	}
-
-	// Handle update: `count++` → `__lazy0[0]++`
-	if (node.type === 'UpdateExpression' && node.argument.type === 'Identifier') {
-		const binding = lazy_bindings.get(node.argument.name);
-		if (binding) {
-			return { ...node, argument: binding.read() };
-		}
-	}
-
-	// Replace lazy variable declaration patterns with generated identifiers
-	if (node.type === 'VariableDeclarator' && node.id?.metadata?.lazy_id) {
-		const lazy_id = create_generated_identifier(node.id.metadata.lazy_id);
-		if (node.id.typeAnnotation) {
-			lazy_id.typeAnnotation = node.id.typeAnnotation;
-		}
-		return {
-			...node,
-			id: lazy_id,
-			init: apply_lazy_transforms(node.init, lazy_bindings),
-		};
-	}
-
-	// Handle identifier references in expression position
-	if (node.type === 'Identifier') {
-		const binding = lazy_bindings.get(node.name);
-		if (binding) {
-			return binding.read();
-		}
-		return node;
-	}
-
-	// Skip JSXIdentifier (component/element names)
-	if (node.type === 'JSXIdentifier') return node;
-
-	// Handle shorthand properties: `{ name }` → `{ name: __lazy0.name }`
-	if (node.type === 'Property' && node.shorthand && node.value?.type === 'Identifier') {
-		const binding = lazy_bindings.get(node.value.name);
-		if (binding) {
-			return {
-				...node,
-				shorthand: false,
-				value: binding.read(),
-			};
-		}
-	}
-
-	// Recurse into child nodes
-	let changed = false;
-	const result = /** @type {any} */ ({});
-
-	for (const key of Object.keys(node)) {
-		if (key === 'loc' || key === 'start' || key === 'end') {
-			result[key] = node[key];
-			continue;
-		}
-
-		// Skip non-computed property keys (they're labels, not references)
-		if (key === 'key' && node.type === 'Property' && !node.computed && !node.shorthand) {
-			result[key] = node[key];
-			continue;
-		}
-
-		// Skip non-computed member expression property names
-		if (key === 'property' && node.type === 'MemberExpression' && !node.computed) {
-			result[key] = node[key];
-			continue;
-		}
-
-		// Skip JSXAttribute name
-		if (key === 'name' && node.type === 'JSXAttribute') {
-			result[key] = node[key];
-			continue;
-		}
-
-		// Skip variable declaration id (the lazy declaration itself was already replaced)
-		if (key === 'id' && node.type === 'VariableDeclarator') {
-			result[key] = node[key];
-			continue;
-		}
-
-		const child = node[key];
-		const transformed = apply_lazy_transforms(child, lazy_bindings);
-		result[key] = transformed;
-		if (transformed !== child) changed = true;
-	}
-
-	return changed ? result : node;
-}
-
-/**
- * Transform default values in function parameters without touching param names.
- * E.g. `(step = count)` where `count` is a lazy binding → `(step = __lazy0[0])`.
- *
- * @param {any} param
- * @param {Map<string, LazyBinding>} lazy_bindings
- * @returns {any}
- */
-function transform_param_defaults(param, lazy_bindings) {
-	if (param?.type === 'AssignmentPattern') {
-		const new_right = apply_lazy_transforms(param.right, lazy_bindings);
-		if (new_right !== param.right) {
-			return { ...param, right: new_right };
-		}
-	}
-	return param;
-}
-
-/**
- * Collect names from a pattern that shadow lazy bindings.
- * @param {any} pattern
- * @param {Map<string, LazyBinding>} lazy_bindings
- * @param {Set<string>} shadowed
- */
-function collect_shadowed_names(pattern, lazy_bindings, shadowed) {
-	if (!pattern || typeof pattern !== 'object') return;
-
-	if (pattern.type === 'Identifier' && lazy_bindings.has(pattern.name)) {
-		shadowed.add(pattern.name);
-		return;
-	}
-
-	if (pattern.type === 'AssignmentPattern') {
-		collect_shadowed_names(pattern.left, lazy_bindings, shadowed);
-		return;
-	}
-
-	if (pattern.type === 'RestElement') {
-		collect_shadowed_names(pattern.argument, lazy_bindings, shadowed);
-		return;
-	}
-
-	if (pattern.type === 'ObjectPattern') {
-		for (const prop of pattern.properties || []) {
-			if (prop.type === 'RestElement') {
-				collect_shadowed_names(prop.argument, lazy_bindings, shadowed);
-			} else {
-				collect_shadowed_names(prop.value, lazy_bindings, shadowed);
-			}
-		}
-		return;
-	}
-
-	if (pattern.type === 'ArrayPattern') {
-		for (const element of pattern.elements || []) {
-			if (element) collect_shadowed_names(element, lazy_bindings, shadowed);
-		}
-	}
-}
-
-/**
- * Collect variable names declared in block-level statements that shadow lazy bindings.
- * Scans VariableDeclarations (const/let/var) and FunctionDeclarations at the top level of a block.
- *
- * @param {any[]} statements
- * @param {Map<string, LazyBinding>} lazy_bindings
- * @returns {Set<string>}
- */
-function collect_block_shadowed_names(statements, lazy_bindings) {
-	/** @type {Set<string>} */
-	const shadowed = new Set();
-	for (const stmt of statements) {
-		if (stmt.type === 'VariableDeclaration') {
-			for (const decl of stmt.declarations) {
-				// Skip lazy destructuring patterns — they ARE the lazy bindings,
-				// not local declarations that shadow them.
-				if (decl.id?.metadata?.lazy_id) continue;
-				if (decl.id) collect_shadowed_names(decl.id, lazy_bindings, shadowed);
-			}
-		} else if (stmt.type === 'FunctionDeclaration' && stmt.id) {
-			if (lazy_bindings.has(stmt.id.name)) {
-				shadowed.add(stmt.id.name);
-			}
-		}
-	}
-	return shadowed;
-}
-
-/**
- * Create a new lazy_bindings map with the shadowed names removed.
- *
- * @param {Map<string, LazyBinding>} lazy_bindings
- * @param {Set<string>} shadowed
- * @returns {Map<string, LazyBinding>}
- */
-function remove_shadowed(lazy_bindings, shadowed) {
-	const result = new Map(lazy_bindings);
-	for (const name of shadowed) {
-		result.delete(name);
-	}
-	return result;
-}
-
-/**
- * Replace lazy parameter patterns with their generated identifiers.
- * A param `&{name, age}: Props` becomes `__lazy0: Props`.
- *
- * @param {any[]} params
- * @returns {any[]}
- */
-function replace_lazy_params(params) {
-	return params.map((param) => {
-		const pattern = param.type === 'AssignmentPattern' ? param.left : param;
-
-		if (
-			(pattern.type === 'ObjectPattern' || pattern.type === 'ArrayPattern') &&
-			pattern.lazy &&
-			pattern.metadata?.lazy_id
-		) {
-			const lazy_id = create_generated_identifier(pattern.metadata.lazy_id);
-			if (pattern.typeAnnotation) {
-				lazy_id.typeAnnotation = pattern.typeAnnotation;
-			}
-			if (param.type === 'AssignmentPattern') {
-				return { ...param, left: lazy_id };
-			}
-			return lazy_id;
-		}
-
-		return param;
-	});
 }
 
 /**
